+++ src/main/java/com/htwberlin/productservice/config/RedisCache.java
@@ -1,4 +1,4 @@
-package com.htwberlin.productservice.config.RedisCache;
+package com.htwberlin.productservice.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
